Fix refetching and initial data handling in course editing
