+++ apps/platform/src/lib/infrastructure/client/pages/workspace/edit/edit-course-general.tsx
@@ -4,13 +4,71 @@ import { CourseImageUploadState } from '../../common/hooks/use-course-image-uplo
 import {
     CourseDetailsState,
     CourseForm,
+    CourseGeneralInformationView,
     DefaultError,
     DefaultLoading,
+    SectionHeading,
 } from '@maany_shr/e-class-ui-kit';
 import React, { useEffect, useState } from 'react';
-import { trpc } from '../../../trpc/client';
-import { fileMetadata, viewModels } from '@maany_shr/e-class-models';
-import { useGetEnrolledCourseDetailsPresenter } from '../../../hooks/use-enrolled-course-details-presenter';
+import { useCourseDetails } from './hooks/edit-details-hook';
+import { fileMetadata } from '@maany_shr/e-class-models';
+
+export function EditCourseGeneralPreview({ slug }: { slug: string }) {
+    const courseViewModel = useCourseDetails(slug);
+    const locale = useLocale() as TLocale;
+
+    if (!courseViewModel) {
+        return <DefaultLoading locale={locale} />;
+    }
+
+    if (courseViewModel.mode !== 'default') {
+        return <DefaultError locale={locale} />;
+    }
+
+    const course = courseViewModel.data;
+
+    return (
+        <div className="flex flex-col space-y-4">
+            <SectionHeading text={course.title} />
+            <CourseGeneralInformationView
+                // These fields aren't utilized and are coming from a common model
+                title={''}
+                description={''}
+                showProgress={false}
+                language={{
+                    name: '',
+                    code: '',
+                }}
+                pricing={{
+                    fullPrice: 0,
+                    partialPrice: 0,
+                    currency: '',
+                }}
+                locale={locale}
+                longDescription={course.description}
+                duration={{
+                    video: course.duration.video ?? 0,
+                    coaching: course.duration.coaching ?? 0,
+                    selfStudy: course.duration.selfStudy ?? 0,
+                }}
+                rating={course.author.averageRating}
+                author={{
+                    name: course.author.name + ' ' + course.author.surname,
+                    image: course.author.avatarUrl ?? '',
+                }}
+                imageUrl={course.imageFile?.downloadUrl ?? ''}
+                students={course.students.map((student) => ({
+                    name: student.name,
+                    avatarUrl: student.avatarUrl ?? '',
+                }))}
+                totalStudentCount={course.studentCount}
+                onClickAuthor={() => {
+                    // Don't handle author click
+                }}
+            />
+        </div>
+    );
+}
 
 interface EditCourseGeneralProps {
     slug: string;
@@ -18,39 +76,32 @@ interface EditCourseGeneralProps {
     setCourseVersion: React.Dispatch<React.SetStateAction<number | null>>;
     courseForm: CourseDetailsState;
     uploadImage: CourseImageUploadState;
+    setIsEdited: React.Dispatch<React.SetStateAction<boolean>>;
 }
 
 export default function EditCourseGeneral(props: EditCourseGeneralProps) {
     const locale = useLocale() as TLocale;
 
-    const [courseResponse] = trpc.getEnrolledCourseDetails.useSuspenseQuery(
-        {
-            courseSlug: props.slug,
-        },
-        {
-            refetchOnWindowFocus: false,
-            refetchOnReconnect: false,
-            refetchOnMount: false,
-            retry: false,
-            staleTime: Infinity,
-        },
-    );
-    const [courseViewModel, setCourseViewModel] = useState<
-        viewModels.TEnrolledCourseDetailsViewModel | undefined
-    >(undefined);
-    const { presenter: coursePresenter } =
-        useGetEnrolledCourseDetailsPresenter(setCourseViewModel);
-    coursePresenter.present(courseResponse, courseViewModel);
-
+    const courseViewModel = useCourseDetails(props.slug);
     const [isFormLoading, setIsFormLoading] = useState(true);
 
     useEffect(() => {
-        if (!courseViewModel) return;
-        props.courseForm.setCourseTitle(courseViewModel.data.title);
-        props.courseForm.parseDescription(courseViewModel.data.description);
-        props.courseForm.setDuration(courseViewModel.data.duration);
-        props.setCourseVersion(courseViewModel.data.courseVersion);
+        if (!courseViewModel || courseViewModel.mode !== 'default') return;
+        const course = courseViewModel.data;
+        props.courseForm.setCourseTitle(course.title);
+        props.courseForm.parseDescription(course.description);
+        props.courseForm.setDuration(course.duration?.coaching ?? undefined);
+        props.setCourseVersion(course.courseVersion);
+        if (course.imageFile) {
+            props.uploadImage.handleUploadComplete({
+                ...course.imageFile,
+                status: 'available',
+                url: course.imageFile.downloadUrl,
+                thumbnailUrl: course.imageFile.downloadUrl,
+            });
+        }
         setIsFormLoading(false);
+        props.setIsEdited(false);
     }, [courseViewModel]);
 
     if (!courseViewModel || isFormLoading) {
