+++ apps/platform/src/lib/infrastructure/client/pages/workspace/edit/hooks/edit-details-hook.ts
@@ -1,0 +1,27 @@
+import { useState } from 'react';
+import { trpc } from '../../../../trpc/client';
+import { viewModels } from '@maany_shr/e-class-models';
+import { useGetEnrolledCourseDetailsPresenter } from '../../../../hooks/use-enrolled-course-details-presenter';
+
+export function useCourseDetails(slug: string) {
+    const [courseResponse] = trpc.getEnrolledCourseDetails.useSuspenseQuery(
+        {
+            courseSlug: slug,
+        },
+        {
+            refetchOnWindowFocus: false,
+            refetchOnReconnect: false,
+            refetchOnMount: true,
+            retry: false,
+            // staleTime: Infinity,
+        },
+    );
+    const [courseViewModel, setCourseViewModel] = useState<
+        viewModels.TEnrolledCourseDetailsViewModel | undefined
+    >(undefined);
+    const { presenter: coursePresenter } =
+        useGetEnrolledCourseDetailsPresenter(setCourseViewModel);
+    coursePresenter.present(courseResponse, courseViewModel);
+
+    return courseViewModel;
+}
