+++ apps/platform/src/lib/infrastructure/client/pages/workspace/edit/hooks/edit-structure-hooks.ts
@@ -15,9 +15,8 @@ export function useCourseStructure(slug: string) {
         {
             refetchOnWindowFocus: false,
             refetchOnReconnect: false,
-            refetchOnMount: false,
+            refetchOnMount: true,
             retry: false,
-            staleTime: Infinity,
         },
     );
